Prepare Navatala GPU public alpha release
Refresh the generated public runtime, kernels, Python bindings, benchmark fixtures, and release documentation for the alpha release.

Add public-surface hygiene checks that reject private repository mentions and publish runtime macros/env vars under the NAVATALA_GPU prefix.

+++ kernels/opencl/cfd/dot_partials.cl
@@ -15,7 +15,7 @@
 
 __kernel void navatala_cfd_dot_partials(__global const float* a, __global const float* b, __global const int* counts, __global float* outPartials) {
   int gid0 = (int)get_global_id(0);
-  __local double tmp[256];
+  __local float tmp[256];
   float v = as_float(0x00000000u);
   if ((((int)((int)(get_global_id(0)))) < counts[0])) {
     v = (a[((int)((int)(get_global_id(0))))] * b[((int)((int)(get_global_id(0))))]);
